requirements.txt
celery==5.3.6
django-celery-beat==2.5.0
redis==5.0.1
flower==2.0.1